Changed The Launcher Icon And Tested On Real Device

+++ lib/src/common_widgets/LineChart/line_model.dart
@@ -1,6 +1,6 @@
 class LineModel {
   final int session;
-  final int accuracy;
+  final double accuracy;
 
   LineModel(this.session, this.accuracy);
 }
